@@ -19,8 +19,8 @@
     "GPL-2.0-or-later"
   ],
   "require": {
-    "php": "^7.2",
-    "typo3/cms-core": "^9.5 || ^10.4"
+    "php": "^7.4",
+    "typo3/cms-core": "^11.2"
   },
   "replace": {
     "typo3-ter/pwd-security-check": "self.version"
